Added middleware to the routes

+++ routes.php
@@ -3,15 +3,15 @@
 $router->get('/', 'HomeController@index');
 
 $router->get('/listings', 'ListingController@index');
-$router->post('/listings', 'ListingController@store');
-$router->get('/listings/create', 'ListingController@create');
+$router->post('/listings', 'ListingController@store', ['auth']);
+$router->get('/listings/create', 'ListingController@create', ['auth']);
 $router->get('/listings/{id}', 'ListingController@show');
-$router->get('/listings/edit/{id}', 'ListingController@edit');
-$router->put('/listings/{id}', 'ListingController@update');
-$router->delete('/listings/{id}', 'ListingController@destroy');
+$router->get('/listings/edit/{id}', 'ListingController@edit', ['auth']);
+$router->put('/listings/{id}', 'ListingController@update', ['auth']);
+$router->delete('/listings/{id}', 'ListingController@destroy', ['auth']);
 
-$router->get('/auth/register', 'UserController@create');
-$router->post('/auth/register', 'UserController@store');
-$router->get('/auth/login', 'UserController@login');
-$router->post('/auth/login', 'UserController@authenticate');
-$router->post('/auth/logout', 'UserController@logout');
+$router->get('/auth/register', 'UserController@create', ['guest']);
+$router->get('/auth/login', 'UserController@login', ['guest']);
+$router->post('/auth/register', 'UserController@store', ['guest']);
+$router->post('/auth/login', 'UserController@authenticate', ['guest']);
+$router->post('/auth/logout', 'UserController@logout', ['auth']);
